requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-filter
PyJWT
sqlparse
python-dotenv
psycopg2-binary
pytz